Fix the profile picture styles
Make it square, so that we get a proper circle of it.

Add the height and width parameters to the img tag. The final size might be different on some clients, but it doesn't matter much. The css will override it if needed.

Make it just a bit larger. Small enough change that it wont make the header any larger.